@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Header from "./Header";
 import Note from "./Note";
 import Footer from "./Footer";
@@ -6,8 +7,18 @@ import CreateArea from "./CreateArea";
 
 function App() {
   var [notes, setnotes] = useState([]);
+  let user = JSON.parse(sessionStorage.getItem('token'));
+  const token = user.token;
+  const defheader = {
+    "Content-Type": "application/json" ,
+    "Authorization" : `Bearer ${token}`
+    };
   useEffect(() => {
-    fetch("/getall")
+    const requestOptions = {
+      method: "GET",
+      headers: defheader
+    };
+    fetch("http://localhost:4000/api/notes/user/getall",requestOptions)
       .then((res) => res.json())
       .then((jsondata) => setnotes(jsondata))
       .catch((error) => {
@@ -17,20 +28,20 @@ function App() {
   function addNote(newNote) {
     const requestOptions = {
       method: "POST",
-      headers: { "Content-Type": "application/json" },
+      headers: defheader,
       body: JSON.stringify(newNote),
     };
-    fetch("/addnote", requestOptions)
+    fetch("http://localhost:4000/api/notes/user/addnote", requestOptions)
       .then((res) => res.json())
       .then((jsondata) => setnotes(jsondata));
   }
   function deleteNote(id) {
     const requestOptions = {
       method: "POST",
-      headers: { "Content-Type": "application/json" },
+      headers: defheader,
       body: JSON.stringify({ id }),
     };
-    fetch("/deletenote", requestOptions)
+    fetch("http://localhost:4000/api/notes/user/deletenote", requestOptions)
       .then((res) => res.json())
       .then((jsondata) => setnotes(jsondata));
   }
@@ -40,8 +51,8 @@ function App() {
       <Header />
       <CreateArea add={addNote} />
       {notes.map((i, index) => {
-        return (
-          <Note
+          return (
+           <Note
             key={i.Ntitle+index}
             id={i._id}
             NoteTitle={i.Ntitle}
@@ -49,7 +60,7 @@ function App() {
             del={deleteNote}
           />
         );
-      })}
+        })}
       <Footer />
     </div>
   );
